refactor(api): split assistant routes and helpers
