+++ tools/size_baseline.txt
@@ -8,7 +8,6 @@ src/quodeq/_cli_evaluation.py:467:function
 src/quodeq/_cli_evaluation.py:574:function
 src/quodeq/_cli_resolution.py:1:file
 src/quodeq/_cli_resolution.py:253:function
-src/quodeq/api/_assistant_helpers.py:190:function
 src/quodeq/api/_evaluation_routes.py:103:function
 src/quodeq/api/_evaluation_routes.py:130:function
 src/quodeq/api/_evaluation_routes.py:183:function
@@ -21,9 +20,8 @@ src/quodeq/api/_run_event_stream.py:1:file
 src/quodeq/api/_run_event_stream.py:289:function
 src/quodeq/api/_sse_log_helpers.py:42:function
 src/quodeq/api/app.py:74:function
-src/quodeq/api/assistant_routes.py:1:file
-src/quodeq/api/assistant_routes.py:239:function
-src/quodeq/api/assistant_routes.py:354:function
+src/quodeq/api/assistant_session_routes.py:58:function
+src/quodeq/api/assistant_turn_routes.py:95:function
 src/quodeq/api/assistant_workspace_routes.py:25:function
 src/quodeq/api/llm_bridge_routes.py:54:function
 src/quodeq/api/routes_findings.py:90:function
